Big commit again
Islands are now generated based on specific seeds. The island selection mode starts with a random seed. The buttons for next and previous changes the seed by +/- 1. There is also the possibility to (temporarily) mark islands as favorites and buttons to jump between them.
Saving last camera position for each island in island selection mode.
Preserving camera position when going from island selection mode to creative mode.
Made sure not to place things with the active tool when the cursor is over GUI.
Tools are now disabled on right click.
The street tool now supports click and drag, with path finding.
Changed the placeholder for bushes to cubes instead of spheres.
New tool for placing boulders.
Added an in game menu and added support for hotkey groups so that other hotkeys stop working when the in game menu is visible.
Added controls for game speed (TimeScaleManager). Work in progress. Speeds 0, 1, 2, 4, 8 and 16. Temporary hotkeys + and -, but not implemented in a good way and don't affect the GUI.